Separating the api from bus consumer worker
